requirements.txt
beautifulsoup4==4.12.3 
cachetools==5.5.0 
fastapi==0.115.0 
langchain==0.3.0
langchain-community==0.3.0
langchain-together==0.1.0
newsapi-python==0.2.7 
pandas==2.2.3 
pymongo==4.8.0 
python-dotenv==1.0.1 
pyarrow==17.0.0 
pydantic==2.9.2 
requests==2.32.3 
streamlit==1.39.0 
tenacity==8.5.0
uvicorn==0.30.6

transformers==4.44.2
torch==2.4.1